Adds working create forgotten password link
